@@ -1,6 +1,6 @@
 package br.com.zupfilms.server.remote;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
