@@ -2,7 +2,7 @@ package br.com.zupfilms.ui;
 
 import android.content.Context;
 import android.net.ConnectivityManager;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
