@@ -1,8 +1,8 @@
 package br.com.zupfilms.ui.home.adapters;
 
-import android.arch.paging.PagedList;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
+import androidx.paging.PagedList;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.FrameLayout;
