@@ -1,9 +1,9 @@
 package br.com.zupfilms.ui.home.adapters;
 
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 import br.com.zupfilms.server.response.FilmGenres;
 import br.com.zupfilms.ui.home.fragments.movieList.MovieListFragment;
