@@ -1,6 +1,6 @@
 package br.com.zupfilms.ui.home.fragments.favorite;
 
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
