@@ -1,7 +1,7 @@
 package br.com.zupfilms.ui.home.fragments.home;
 
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
+import com.google.android.material.tabs.TabLayout;
+import androidx.viewpager.widget.ViewPager;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
