@@ -1,9 +1,9 @@
 package br.com.zupfilms.ui.home.fragments.home;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.Observer;
-import android.support.annotation.Nullable;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.Observer;
+import androidx.annotation.Nullable;
 
 import br.com.zupfilms.model.ResponseModel;
 import br.com.zupfilms.server.response.FilmGenres;
