@@ -1,6 +1,6 @@
 package br.com.zupfilms.ui.home.fragments.movieList;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
