@@ -1,7 +1,7 @@
 package br.com.zupfilms.ui.home.fragments.search;
 
-import android.support.design.widget.BottomNavigationView;
-import android.support.v7.widget.RecyclerView;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.FrameLayout;
 import android.widget.ProgressBar;
