@@ -2,9 +2,9 @@ package br.com.zupfilms.ui.home.homeActivity;
 
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.BottomNavigationView;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import androidx.fragment.app.Fragment;
 import android.text.SpannableString;
 import android.text.style.StyleSpan;
 import android.view.MenuItem;
